refactor(recipes): relocate dietary-profile provider as consumer-owned contract
Final cross-module contract migration. Recipes is the only consumer, so
the contract and DTO move into Recipes.Application; the HTTP adapter
takes a Recipes.Domain.Recipe.OwnerIdentifier.

- new Recipes.Application.Interfaces.IDietaryProfileProvider (OwnerIdentifier VO)
- new Recipes.Application.DTOs.DietaryProfileSnapshot
- HttpDietaryProfileProvider relocated to Recipes.Infrastructure/ExternalServices
- removed Shared.Common.IDietaryProfileProvider + DietaryProfileSnapshot
- GetRecipeSuggestionsQueryHandler now passes currentUser.AsOwner()
- removed obsolete Recipes.Infrastructure/Services/ folder (now empty)

Yumney.Shared.Common no longer hosts any cross-module service interfaces;
it's left with primitive cross-cutting types only, matching the rule
documented in CLAUDE.md (#527).
